Update a system to support a lot of connection
